pyproject.toml
[project]
name = "kevin"
version = "0.1.0"
description = "AI-powered call screening assistant"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "twilio>=9.3.0",
    "httpx>=0.27.0",
    "python-telegram-bot>=21.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "phonenumbers>=8.13.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "google-cloud-firestore>=2.19.0",
    "firebase-admin>=6.5.0",
    "google-cloud-secret-manager>=2.21.0",
    "google-cloud-tasks>=2.17.0",
    "google-cloud-logging>=3.11.0",
    "slowapi>=0.1.9",
    "websockets>=13.0",
    "PyJWT[crypto]>=2.9.0",
    "httpx[http2]>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "respx>=0.22.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
